Changes to enable writing/reading of refined meshes using a updated XDA file format that stores adaptivitiy hierarchy information.
The main changes are in xdr_io.C, in which the new XDA format is written
out in write_mesh, and adaptive meshes can be reconstructed in read_mesh.
These functions should still work properly for old-style libMesh and MGF
XDA files.

A new version of mesh_communication.C has been checked in, which
broadcasts and reconstructs adaptive mesh information, although this
version does not appear to be working properly yet... when reading in
adaptive meshes in parallel, DOFs can be mixed up giving nonsense solutions.
This file needs to be looked at...