Adding the Variational Smoother class.
This checkin adds the Variational Smoother class.  The original
smoothing/redistribution code was written by Larisa Branets at the
University of Texas at Austin.  The code was then modified to work
within the libMesh framework by myself.

There are three constructors to use depending upon your
purpose... each one more complex than the next.  There is one just for
smoothing, one for redistribution, and one for redistribution using an
area_of_interest mesh.

The first acts a lot like a modified Laplace smoother... just
increasing element quality.  Defaults are given for many of the
parameters, these defaults should be good enough for general smoothing
purposes.

The second two can smooth and redistribute the mesh simultaneously.
The redistribution happens by passing a vector of values (usually
corresponding to an error indicator, in fact ErrorVectors can be
passed directly).  This vector of values is then used to weight the
functional during the optimization process.  Again, sane defaults are
provided for many variables... but for good redistribution these will
probably need to be modified.

Better documentation for all of this is forthcoming... along with a
few tests to excersize the capabilities.

Derek


git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@1907 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf