The DenseMatrix::lu_solve() routine now has only one implementation, and that implementation uses partial pivoting.
We have also enforced the notion that lu_solve should
not modify the rhs entry passed to it by making the
reference passed a const one.

Fixed up the det() routine to handle the number of
row interchanges in the LU decomposition and compute
the proper sign of the determinant.