The DenseMatrix::lu_solve() routine now has only one implementation, and that implementation uses partial pivoting.
We have also enforced the notion that lu_solve should
not modify the rhs entry passed to it by making the
reference passed a const one.

Fixed up the det() routine to handle the number of
row interchanges in the LU decomposition and compute
the proper sign of the determinant.



git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@3588 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf