Added new utility header and (empty) source file for the "hashword" implementation by Bob Jenkins.
Replaced compute_key implementations with hashword calls.

All hashing functions are inlined to help with speed.

Hardcoded hashing function for exactly 2 ints added.

Anonymous and Utility:: namespaces used to avoid polluting.

The new hash function can only be used if we can
guarantee that unsigned ints are 32 bits, since the hash
function assumes this.

The code falls back on the old hash if this isn't the case,
I'd be very surprised if that happens any time soon, or ever.