Working on a system for generating and using constraint data that's indexed by nodes, not just by dof indices.
In the long run this is necessary for us to support snapping nodes to
curved surfaces in adaptive 3D runs.  In the medium run, I think I can
make compute_sparsity much faster in heavily-multiphysics (~10
variable) cases.

In the short run this is going to be a series of fairly significant
changes that need testing and won't pay off right away, so I'm putting
it in a branch.



git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@4660 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf