Use libtool in Make.common as well - seems to help fix Moose issues.
Requires installing the libtool used to build libMesh, for platforms
that do not have an adequate libtool of their own.  Hopefully this
allows MOOSE to work with the latest libMesh.

Note that applications that are using Make.common to define their
build rules might need to augment their clean rules.