Moving mesh stuff moved to DiffPhysics
We still need FEMPhysics for the eulerian_residual default
implementation, but there's nothing about moving meshes in general
that's FEM specific, so let's move the moving mesh variables from
FEMPhysics down to DiffPhysics.

We should do the same with FEMContext->DiffContext, for the same
aesthetic reasons, but I'm doing this move first since it also rids us
of an otherwise unavoidable dynamic_cast and thus an RTTI dependency.