Use explicit init.comm in examples
This should let "make check" pass with --disable-default-comm
eventually.  For now, we break when we get to the adaptivity examples
because there's nothing with a Communicator initializing the
ErrorEstimator subclasses.  I think I want to fix that one with some
compatibility work on ErrorEstimator, though, not by requiring more
changes to user code.