GCC 4.6.3 does support -std=c++0x, but it doesn't support override.
So make the new configure test a bit less stringent.  This test is a
little tricky to get right -- it has to have *some* C++11 stuff to
accurately determine what the correct flag for turning on C++11
support is, but it can't test lots of features otherwise C++11 will
sometimes be disabled completely without good reason.