Some opt-mode error checking in Parsed*Function
Leaving these tests on shouldn't be too much overhead relative to
the parsing process that already takes place in these constructors.

And since the whole point of parsed functions is to allow users to
change the string contents at run time, there's no way for us to
guarantee that those strings will be valid the way we can (attempt to)
make such guarantees about other assertions.