Convert LOC plot from Octave to Matplotlib
Support for GNU Octave in OSX seems to be a low priority.  In one of
their bugzillas, it actually says "The Octave developers are very
short on Mac OS X resources."  These plots may be easier to maintain
as numpy/matplotlib scripts instead...

This is pure python-coding-by-google, so it is probably terrible, but
it gets the job done.