Comment out overzealous assertions
When we hit set_node_processor_ids() in a skip_partitioning()
codepath after doing enough coarsening, we may legitimately be in a
state where orphaned nodes still exist but don't still have valid
processor ids.

The best thing to do in the long term may be to reorder the deletion
of orphan nodes so as to avoid this state, but that's tricky to do,
so for now we'll just remove the assertions that are giving us
false-positives.