Bug fix for PointLocatorList + unit test
Before, the cached distance used the distance from the first element
in the list. But, the loop didn't include the first element so if
the first element was the closest one, we didn't detect it and the
accompanying unit test failed. I put the unit test in the mixed
dimension unit test because this bug was found as part of some
refactoring I'm doing and I went ahead and made this bug fix standalone.