Adding new API for mixed dimension support in PointLocators
Old API is preserved for backward compatibility. Note that the behavior
of the PointLocators hasn't changed yet, just the API. I have at least
included checks on the element dimension so if someone manages to use
this commit, they'll still get an error. Note I put the old API
implementation in the source instead of the header to avoid including
mesh_base.h in the PointLocatorBase header.