Ran contrib/bin/reindent.sh on headers.
Note that the LIBMESH_ENABLE_TOPOLOGY_CACHES macro was confusing
Emacs' indenting algorithm.  I fixed this by adding a semi-colon to
the end of the macro invocation.  Apparently semi-colons are
explicitly allowed at the end of a function definition in a class, so
this fixes the indentation issue and is also legal C++.

http://stackoverflow.com/questions/5365604/semicolons-in-a-class-definition