Add scoped_lock to Elem methods with static caches
This should fix a possible race condition, if we ever call these
methods from multiple threads.

I'd *like* to be able to set a static "cache_is_completed" bool and
skip the lock after it's true... but without knowing the maximum
possible value for em_vers, there's no way to be sure the cache is
completed.  Adding a "max_embedding_version()" method might be an
option in the future in the unlikely event that these locks become
bottlenecks.