Fix for Intel-15
For some reason, the Intel compiler with a GCC 4.3 STL refused to
convert the constant NULL to pointer-to-NumericVector.  Pretty sure
the compiler is wrong here, but the fix isn't terrible.

We tracked the issue down to compiling with the -std=c++0x flag and
the following very simple test code:

```
int main()
{
  std::vector<int*> ivec;
  ivec.push_back(NULL);

  return 0;
}
```

"Intel 15 with libstdc++ 4.3 sucks".
@roystgnr, 2015