Adding dlopen configure test.
We have been testing for <dlfcn.h> and searching for libraries which
provide dlopen for a while. This test formalizes the process, tests
for dlsym() and dlclose(), and sets the LIBMESH_HAVE_DLOPEN
preprocessor macro.  The fparser JIT test, which used to check for
dlopen, now simply uses the results of this test.