Don't use cpp_compress in fparser
It was already removed from the optimizer code generation, but we
missed it in this opcode code generation.

This fixes a compilation bug with clang 3.5 for me: the definition of
a macro named "q1" here (why not "#define i" while we're at it?!)
conflicted with an argument name in the emmintrin.h header.