Eigen should work again for the reduced basis examples.
Run them with --use-eigen to verify.  I found that Eigen's BICGSTAB
solver gave unreliable results in in reduced_basis_ex5: sometimes
passing, sometimes taking >10000 iterations to converge.  It honestly
feels like a bug, but I don't have time to debug it at the moment.
Luckily Eigen's CG solver can be used in the "offline" mode for this
example (as is done with PETSc) and it works just fine.

This reverts the following commits:
* 994024990385f8a1aa8981b11826376318e80042
* 0140e47b9db39d25f54cf83cda7372ef69b5aef0
* 07a1e83ffc3ba4ad46626b1be2762fa324e9eb9f