Replace uses of "echo -n" in shell scripts with printf.
Technically this is only required for scripts run with /bin/sh, but I
think it's wise to fix all the instances we can find, just to avoid
this issue in the future.

To find all shell scripts that had non-POSIX echo commands, I did:
git ls-files '*.sh' | xargs grep "echo -"

To test this, I ran all the rebuild_XYZ.sh scripts and the associated
files didn't change, so I assume that it worked.

Refs #575.