Handle first_solve in NewmarkSolver
Initial displacement and velocity are given, but we also need the
initial acceleration. This is done by solving the system of equations
for the initial acceleration. However, the system is otherwise posed
for solving for the updated displacment (from which the updated
velocity and acceleration can be computed). So, to accomodate this,
we override UnsteadySolver::solve and if we are on the first solve,
we swap the solution with the acceleration vector and then assemble
accordingly in the NewmarkSolver::_general_residual.