Fix sign error
There was a sign error in the reference I was using (*shakes fist*).
I redid the math myself and there was a sign error in the acceleration
term. This fixes both the excessive numerical dissipation I was
seeing as well as "stagnant" solution for computing initial_acceleration
with zero initial velocity and zero initial displacement.