Refactor parallel_algebra.h
We can simplify our derived types slightly by noting that the
LIBMESH_DIM entries are homogenous.

We should wrap MPI calls in the new error checking macro.

We should commit intermediate MPI types; failing to do so is
infinitesimally more efficient and seems to work in practice, but the
error descriptions in the resize docs I've read suggest that using an
uncommitted type may not be strictly allowed.