Explicitly detach from running process when doing GDB backtrace.
Just checking whether this might help with the zombie gdb process
issue described in idaholab/moose#4201.  I'm not sure if this actually
makes a difference, but I did find two different versions of the
following sentence in online gdb docs:

http://www.delorie.com/gnu/docs/gdb/gdb_23.html
"If you exit GDB or use the run command while you have an attached
process, you kill that process."

https://sourceware.org/gdb/onlinedocs/gdb/Attach.html#Attach
"If you exit gdb while you have an attached process, you detach that
process. If you use the run command, you kill that process."

So, it's probably best to add an explicit "detach" just to be on the
safe side...