Add libmesh_override to include/reduced_basis headers.
Note: The curiously recurring template pattern can make it a bit tricky
to track overrides.  For example, the declaration:
```
class TransientRBConstruction : public TransientSystem<RBConstruction>
```
means that RBConstruction is actually a base of
TransientRBConstruction and can provide overrides for its virtual
functions.

Another example is:
```
class RBConstruction : public RBConstructionBase<LinearImplicitSystem>
```
which means that LinearImplicitSystem is a base of RBConstruction.