Deprecate QuadratureRules::name().
This function and the rest of the stuff in the QuadratureRules
namespace have not been kept up-to-date with changes in the rest of
the library, e.g. there are more than 3 valid quadrature rules now.
They don't seem to be used anywhere, and have largely been supplanted
by the Utility::enum_to_string() stuff anyway.

include/quadrature/quadrature_rules.h
src/quadrature/quadrature_rules.C

will be removed in a month or right before the next release, whichever
comes first.