send_recieve: try harder to pick an example T
Some StandardType implementations may *require* an example to
calculate offsets used to build the type.

This is the best we can do from within send_receive().  It's clearly
not good enough in general (we may have ranks whose send vectors are
empty and receive vectors preallocated) but the only other option is
to try and construct an example with a default constructor, which for
some T won't exist.