Avoid clear() in DiffSystem class destructors
The previous code doesn't do what it might be expected to do.  Since
the methods called by clear are virtual, and any subclasses have been
partially-destructed down to the base class at this point, and the
base class implementations of physics and qoi clearing are no-ops,
calling those functions does nothing.  If clearing of subclass data
needs to be done then it needs to be done in the subclasses'
destructor.

On Intel 16 in opt mode, those virtual function calls give me a
segfault, which I'm pretty sure is a compiler bug, but since fixing my
stupid code also avoids the segfault I'm not complaining too hard.