Set interior parent only when mesh contains multiple dimensions
This commit is to address the issues brought up in libMesh/libmesh/#709,
namely, handle ParallelMesh more appropriately by using
LIBMESH_BEST_UNORDERED_MAP and asserting when an element id is greater
than the maximum element id rather than the number of elements.  Also,
automatically setting the interior parent should occur when a mesh has
multiple dimensions and skipped when the mesh has only one dimension.

In order to utilize mesh.elem_dimensions(), which allows a user to
determine the dimensions of a mesh, the code to automatically set the
interior parent was moved into mesh_base.C as a separate method.  This
way mesh.cache_elem_dimensions() can be called prior to setting the
interior parents and thus mesh.elem_dimensions() will be available.

Also, the methods were moved prior to the partitioning in order to avoid
the complexities of one processor containing the interior parent of an
element on another processor.

Lastly, there is one noticeable performance penalty for moving the
automatic setting of the interior parents to a separate method and that
is populating the node_to_elem map which requires iterating through all
active elements.  Previously, the node_to_elem map was populated during
an existing element iteration inside find_neighbors().