Fix DenseMatrix::svd(sigma, U, VT);
Lapack is actually computing the SVD of A^T, which is given by:

A^T = V S U^T

so:

Lapack | Is actually
-----------------
U      | V
V^T    | U^T

Thus we need to not only swap the meaning of U and V before handing
them back to the user, we also need to transpose them, so the user
gets U and V^T as he is expecting.  However, since Lapack already
stores these matrices in column-major order, we get the transpose for
"free" and can achieve the desired effect simply by swapping the order
in which the arguments are passed to the _svd_helper function.