Move Quad4::key() to Quad::key().
Since Hex20 elements compute their side keys using the four corner
vertices of each face, this is also the appropriate way to compute the
key for a Quad8 element.  Moving the Quad4::key() implementation to
the base class allows it to be used by the Quad8 as well.  Note that
the Quad9 will continue to specialize this function.