Move Tri3::key() to Tri::key().
Since e.g. Prism15 elements compute their side keys using the three
corner vertices of the quadratic triangular faces, this is also the
appropriate way to compute the key for a Tri6 element.  Moving the
Tri3::key() implementation to the base class allows it to be used by
the Tri6 as well.