close_to_point needs tol/hmax
Since close_to_point multiplies tolerance by hmax (to account for
errors in master space at a hopefully-appropriate scale) we need to
divide tolerance by hmax/2 (i.e. hmax for a child element) to be sure
we have enough leeway to find a point within that tolerance in
physical space, such as the perturbed points we use to get C1
elements' mixed derivatives.