Add DenseMatrix::svd_solve().
This function fills a missing requirement in the DenseMatrix classes,
allowing us to solve non-square systems of equations in a
least-squares sense.  The user can pass a tolerance to svd_solve()
which determines the cutoff for small singular values. svd_solve() is
a const member function: we make a copy internally instead of allowing
Lapack to modify A.

Note that Eigen also has the capability to solve non-square systems of
equations, but it is relatively slow, as discussed in this thread:
https://forum.kde.org/viewtopic.php?f=74&t=102088, so having our own
Lapack-based implementation is worthwhile.