Add testRestart to slit mesh test
This doesn't currently pass for me with ParallelMesh.

This is a bug with the underlying code, not a bug with the test -
reading the restart files with two processors, the parallel solution
is be split differently than the mesh nodes.

It turns out that the partitioning we have when we write a serial xda
file to disk is not guaranteed to match the partitioning we have when
we read it back in, but our current parallel solution strategy depends
on it matching.