Fix deprecation warning in unit tests.
face_level_mismatch_limit() is supposed to set the value of an
unsigned char, but this line of code was assigning it the value of
"false", which is equivalent to 0, which is equivalent to unlimited
mismatch, which triggered the deprecation warning.  Based on the
nearby comment, we don't want to override any settings anyway, so I
think this is the right fix.