Add Quad8::volume().
This speeds up a test code which calls elem->volume() on each
distorted QUAD9 in a 4M element mesh by approximately 650x (from about
206s down to 0.3146s).

As far as the accuracy of the four-point quadrature rule goes, it
appears to be quite good up to about 15% distorted elements (in the
context of MeshTools::Modification::distort()).  More distortion than
this is probably not really usable for finite element analysis,
anyway.  In the table below, max_rel_diff is computed using a
twelfth-order quadrature rule (49 points) in the reference volume
computation.

distortion max_rel_diff
0.05       3.0166222190035792e-15 (<1%)
0.1        2.9337828206309171e-15 (<1%)
0.15       1.9220118208802948e-02 (2%)
0.2        1.4316734561689504e-01 (14%)
0.3        3.6570789827049105e-01 (36.5%)

It would also be possible to implement an "early return" branch for
this function, but you would have to check the value of 8 different
3-vectors against zero so I didn't think it was worth the extra
complexity for the minimal performance improvement.