Add notion of time-dependence to FunctionBase
FunctionBase objects are assumed to represent functions of (x,y,z,t).
It is convenient to be able to query whether the function is explicitly
a function of time, e.g. for deciding whether or not to reinit
DirichletBoundary constraints. The function is assumed to be
time-dependent unless the user or subclass says otherwise. For
subclasses that cannot natively determine time-dependence, there
is a setter method.