Add Quad9::volume().
This implementation is roughly 350x faster than the base class
Elem::volume() implementation, and also uses 3x3 quadrature, so should
be returning the same results.

The table below gives the maximum relative difference in the volume()
computation on 20x20 meshes of QUAD9s with different distortion
levels.  The difference is relative to a twelfth-order quadrature rule
with 49 points.

distortion  max_rel_diff
0.05        (<0.00%)
0.1         (<0.00%)
0.15        0.73%
0.2         3.67%
0.3         10.95%