Add Edge4::volume().
This speeds up calling volume() on every element in a mesh with 4M
Edge4's by about 450x. Though I'm not sure this was ever going to be a
performance concern, I still want to have a custom volume()
implementation for all the different Elem types for completeness.

The 4-pt quadrature is also reasonably accurate.  The biggest
discrepancies seem to be introduced by perturbing the interior nodes
of the EDGE4. For example, perturbing an interior node of the
reference element by 0.125 in the y (or z) direction leads to a
relative error of much less than 1%, compared to a 12th-order (7-pt)
approximation of the volume.

Approximate volume  = 2.0412301724043291e+00
Verification volume = 2.0410437292661414e+00
Rel err             = 9.1346959163268581e-03%