Optimize Hex8::volume() slightly.
If my calculations are correct, the "geometric" formula I used
previously had about 12 dot products and 12 cross products, while the
current one only has 4 of each.  This formula is derived by writing out the
standard volume formula and dropping terms which are zero due to the
symmetry of the integrand and/or triple products containing two copies
of the same vector.

While the original geometric formula was already pretty fast, this one
is about 1.7x faster (about 0.1697s vs 0.2866s to call volume() on
3.375M Hex8 elements).