Add Prism15::volume().
This speeds up the time required to compute volume() on about 3.9M
Prism15 elements by approximately 245x over the base class
implementation (2.6091s vs. 640.7658s).

Note: we can get away with a 4th-order (rather than 5th-order) 2D
quadrature rule for this element, so that saves a few operations.  The
other big win in this optimization is pre-computing the pow terms and
doing table lookups in the loops.  Without that, the same code takes
about 10x longer to run.