Revert "Add quadrature-based volume() implementation for Pyramid5."
I'm reverting this because the quadrature implementation was actually
about 13x slower than the existing geometric formula.  The geometric
formula is so simple, there's really no way a quadrature-based
implementation, especially with eight quadrature points, is going to
beat it.  Also, since we're dealing with rational basis functions,
there is a division in the quadrature loop, which isn't ideal.
Although I'm reverting this change, I'd like it to stay in the history
so I can refer back to it later.

This reverts commit f1ae44a93140469423921a8e2fcff789e41c2086.