Add Pyramid13::volume().
This speeds up the time required to compute volume() on about 3.68M
Pyramid13 elements by approximately 60x over the base class
implementation (653.67s vs. 11.0274s).

It also unfortunately copy/pastes a bunch of similar code from the
Pyramid14::volume() implementation.  I suppose it could be moved to
the Pyramid base class, but it's not really relevant for the
Pyramid5...