Add support for --with-thread-model=none.
This allows a user to explicitly turn off threading, and should enable
the "dummy" versions of things like spin_mutex, recursive_mutex, etc.
I've also decided to disable OpenMP in this configuration, but I
suppose there could be use cases for still configuring the OpenMP
compiler flags for dependent apps...

Also, OpenMP is not an 'independent' threading model like TBB or
pthreads.  It works in conjunction with the pthread threading model's
parallel_for() and parallel_reduce(), but OpenMP is also useful for
other purposes, and if libmesh properly detects the flags required to
turn on OpenMP, apps which use libmesh can use OpenMP pragmas.

Refs #838.