Add PerfItem class and LOG_SCOPE/LOG_SCOPE_IF macros.
This gets us about 95% of the way to the RAII PerfLog feature
discussed in #199.  It uses the macro trick to generate a unique
variable name based on __LINE__, but instead of outright replacing the
existing START_LOG/STOP_LOG macros with this new macro, I've
introduced a new one.  There are situations where a scope-based
logging macro just doesn't make sense, but they are pretty rare.
Almost all START_LOG/STOP_LOG pairs in the library can be replaced
with the new LOG_SCOPE macro, and the ones that can't easily be
replaced can just remain unchanged.

The logging capability can also be disabled at runtime by passing the
appropriate flag to the LOG_SCOPE_IF macro.  See examples of the way
this is used in the dof_map_constraints.C code.