Opt out of GCC pragmas with Intel-2013.
Otherwise, we were getting an error about missing semi-colons.  It
appears that this version of Intel does not support "_Pragma" which is
how you get a #pragma command to be generated via macro expansion.
The full error message is below:

ncgentab.c(1328): error: expected a ";"
    _Pragma ("GCC diagnostic push") _Pragma ("GCC diagnostic ignored \"-Wuninitialized\"") _Pragma ("GCC diagnostic ignored \"-Wmaybe-uninitialized\"")

According to the GCC docs (https://gcc.gnu.org/onlinedocs/cpp/Pragmas.html),
_Pragma was introduced in C99.  Is it possible that the Intel-2013
compiler is not using -std=c99 for C code, or that it does not
implement the full standard?