Clean up after removing deprecated headers/src.
This includes moving from a deprecated warning message when
reading/writing "legacy" Xdr files to an error.

The meshtool app used to allow you to read in a "legacy" format XDA/R
mesh and solution file, and write it back out in some other format.
This is actually a cool feature, but I don't think anyone has used it
in a while, because we have not really been writing "legacy" XDR/A
files for quite some time.  Left a few lines of instructions about how
it could be accomplished again if desired.