Use (node/elem)_(ref/ptr) API names
As discussed in https://github.com/libMesh/libmesh/issues/889 these
should be easier for users to remember.

The old API names aren't libmesh_deprecated() here, but are noted as
deprecated in comments.  We'll probably add a libmesh_deprecated()
after the next release, and remove the old names a couple releases
later.