Fix for not-quite-compatible boost::unique_ptr.
With std::unique_ptr it is apparently fine to create and return one
from a function, e.g.

std::unique_ptr<int> foo()
{
  std::unique_ptr<int> x(new int(1));
  ...
  return x;
}

but this does not work with boost::unique_ptr, because the compiler
tries to call the private copy constructor.  This workaround simply
releases control from the existing UniquePtr before immediately
handing it over to the control of a UniquePtr created in the return
statement.