Rename SerialMesh and ParallelMesh
This is one big fat commit, thanks to chicken-and-egg problems.

We repeat the same trick we did once upon a time with Mesh: use a
subclass rather than a typedef for our backwards compatibility shim,
so as to avoid breaking user code forward declarations.

This could cause breakage for anyone who's templating code around mesh
type.  Ironically, *we* seem to be templating code around mesh type in
the MeshOutput class, but that appears to be an atavism and we don't
actually use anything but MeshOutput<MeshBase>.