Fix multiple statements on single line.
IMO this makes the code harder to read for humans, harder to set
breakpoints for in debuggers, and has no real benefits. Also, the way
that some of the if-statements were formatted made it appear that the
iterator increment was intended to be *inside* the preceding
if-statement, but the increments are supposed to happen no matter
what. Also fix small whitespace issue.