Update to MeshFunction::clone()
Call clone->init() inside the clone method if the original
object was initialized. This is important when you want to
use a MeshFunction with System::project(), because project
makes a clone of the MeshFunction and we need to make sure
that the clone is initialized.

Also added a new unit test to using a MeshFunction with System::project().