Remove overzealous asserts in FEMContext
When we have the case of subdomain-restricted variables, the dof_indices
only get populated with indices from the variables currently active
on the element. Thus, if the variable var for which we are querying
was inserted in the System after variables that are not active on
the current element, then this assert will fail. The more appropriate
assert actually happens in the get_dof_indices(var) call, so we
just remove these asserts entirely.