Fix const-correctness issue in BoundaryInfo.
We have always held a const reference to a MeshBase in BoundaryInfo,
but the sync() function needs to perform actions on the Mesh which are
technically non-const.  It builds non-const sides from Elems in the
Mesh for the purpose of inserting them into the BoundaryMesh, but this
is no longer allowed by the new const-correct build_side_ptr().

There are a couple of potential fixes for this situation, including
const_cast, but I think the right solution is to make the MeshBase
reference non-const, as this more accurately reflects the way in which
the Mesh is used by the BoundaryInfo object.  _mesh is a private
parameter and there are no public accessors, so this change does not
risk providing non-const access to clients of the BoundaryInfo object.