Fix const-correctness issue in InfFE::Base::build_elem().
This function is required to produce non-const Elems from incoming
const Elems by calling build_side().  The constness of the incoming
Elem is required by other APIs, specifically FEAbstract::reinit(),
which should not be changed.  This function was always doing something
a little fishy, it is just being made explicit with a const_cast now.