Handle NULL elements in EquationSystems::get_solution().
Previously, we were using:

mesh.n_elem()

to determine the global size of the parallel_soln vector, and

std::distance(_mesh.local_elements_begin(), _mesh.local_elements_end())

to determine its local size on each processor.  These two sets of
numbers are not consistent when there are NULL Elems in the Mesh
because the former counts the NULL entries while the latter does not.
This would happen if e.g. get_solution() was called prior to calling
MeshBase::contract().  The fix is to just compute the local sizes
given the global size, without relying on Mesh iterators.  Values for
contracted elements simply show up as zeros in the solution vector
which gets built, it is up to the individual I/O formats to handle
this correctly (Exodus should be able to).