Call libmesh_convergence_failure() from TypeTensor::solve,inverse().
Note 1: TypeTensor::solve() used to only check for zero determinant in
debug mode.  In optimized mode, it would go ahead and divide by zero.
This is not a good approach, not only because it behaves differently
in optimized vs. debug mode, but also because the optimized code path
may raise a floating exception on division by zero, depending on the
system configuration.

The current approach is to throw an exception when a floating point
zero determinant is encountered regardless of optimized/debug mode,
and let the calling code handle the exception appropriately.

Note 2: libmesh_convergence_failure() doesn't seem like it will
compile if !LIBMESH_ENABLE_EXCEPTIONS, since in that case the
ConvergenceFailure class does not even exist.  We may not have tested
the no exceptions configuration recently...

Note 3: This isn't actually a convergence failure, so if we decide to
go this route, we may want to define one more custom class derived
from std::runtime_error() which has an empty definition when
!LIBMESH_ENABLE_EXCEPTIONS.