Fixes for --disable-exceptions.
I thought there would be more issues, but the LIBMESH_THROW macro is
very clever -- when exceptions are not enabled, the argument to
LIBMESH_THROW is not used for anything, so e.g.

LIBMESH_THROW(libMesh::ConvergenceFailure());

is fine even though libMesh::ConvergenceFailure is not even defined
when exceptions are disabled.

Refs #1034.