Remove DenseMatrix::svd()
I added svd to DenseMatrix a few years ago. On reflection, I prefer to remove it
since I've found that the current implementation fails in some cases. The LAPACK
interface that is used under the hood is complex and has many options, and rather
than try to handle all those options correctly in all cases, I prefer to just
remove this method. If users want an SVD of a DenseMatrix, it's better if they
call LAPACK themselves (which they can do via the PETSc interface to LAPACK).