Add DenseMatrix routines to compute left and right eigenvectors.
Note: if user requests eigenvectors, we must transpose A in-place
before calling LAPACK, otherwise LAPACK will return the
inverse-transpose of the matrix we actually want.  Since this in-place
transposition isn't free, we only do it when eigenvectors are
requested (it's not necessary when only eigenvalues are being
computed).  Also note that this routine has always overwritten the
matrix; doing this in-place transposition does not introduce any
additional side-effects.

Refs #1059.