Add unit test verifying left and right eigenvector calculation.
I had to loosen the tolerance on the eigenvalue checks slightly, since
one of them just barely failed after I introduced the transposition of
the matrix.  There does not seem to be a tolerance to control this in
LAPACK, this just seems to be the accuracy you get for the test
problems.

Refs #1059.