Add unit tests for Euler(2)Solver
This uses FEMSystem to implement simple ODEs to integrate in time
and unit test the Euler solvers. In particular, we have one that
should be integrated exactly for first order solvers and one
that should be integrated exactly with second order solvers.

There are currently two issues I've found here.
1: Using a SCALAR
variable (and nonlinear_time_derivative and nonlinear_mass_residual)
gives the *WRONG* answer. I'm not sure why yet - this needs to be
investigated further.
2: Euler2Solver does not integrate exactly the LinearTimeFirstOrderODE
System. I've found the problem and will fix it subsequently.