Add initial reinit_func(1.) call in Euler2Solver
We need the call to reinit_func to set the time, t, in the context
to the correct value. Also added clarifying comments in EulerSolver
and Euler2Solver that we're also setting the time in addition to
possibly resetting the mesh if there's mesh motion.

There is probably a way to make this more efficient such that we
only call reinit_func twice in Euler2Solver, but I didn't put any
thought into it.