Add BoundaryVolumeSolutionTransfer class.
This class can be used for transferring solutions between the surface
of a volume mesh and the BoundaryMesh associated with that surface.
This is joint work with with Xikai Jiang from Argonne National
Laboratory (@xikaij) who is the original author of the transfer.

See also: X. Jiang et al., "An O(N) and parallel approach to integral
problems by a kernel-independent fast multipole method: Application to
polarization and magnetization of interacting particles," The Journal
of Chemical Physics vol. 145, 064307, http://dx.doi.org/10.1063/1.4960436.