GhostingFunctor needs 2 reinit options
We need a mesh_reinit() so we can update internal caches ASAP when the
mesh changes, before it's redistributed or remote elements are
deleted.

But we also may need a dofmap_reinit() for functors which do
non-default things with their CouplingMatrix values and which can't do
that properly until variables, dof indexing, etc. are all in order.

So let's provide both and let users decide what (if anything) they
need to override.