Add evaluable_elements_begin()/end()
I put these in MeshBase like all the other iterator ranges, for
simplicity and consistency, but we do still need a DofMap reference to
initialize them.

Iterating through these might be slow, but creating a correct mutable
cache instead will definitely be a huge pain, so let's just get the
code correct for now and optimize later if we need to.