assertion for the MPI1-with-32-bit-long case
Not sure if it's safe to put a static_assert here too, so I decided
not to.  In theory anyone in this use case can turn all their id types
down to 32-bit and still have a working libMesh, and we  shouldn't
risk breaking that.

If you're using MPI 1.0, Windows, and only 32-bit integers, I don't
want to risk making your life any more difficult than it already is.