Update traffic.
There seems to be a bug in either the script that we use to download
the statistics, or in the web page data itself, that affects the first
day's results.  I confirmed this by checking the stats on both
Wednesday and Monday, and both times the stats for the first day of
the range returned were affected, but no other days were.  Visiting
the website also confirmed that the first day's data was not actually
present on the graph itself.  So I'm not sure where the issue lies,
but a simple fix seems to be to ignore the data from the first day
returned by the script.