Make send values const
In the overloads where we're specifying containers as send values,
this is necessary for correctness; users really might want to send a
const container.

In the "typename T" cases, this isn't necessary for correctness, but
it's good for being self-documenting.  The overhead of passing a small
POD by const-reference instead of value should be negligible in debug
mode and zero (since these functions are all inline) in opt mode.