Remove all XYZ caching
This is definitely a bug.  We could reintroduce a cache someday as
member variables of an FE subclass, but using static variables doesn't
just break in the case of threads: it can break with XYZ elements on
multiple meshes, it can break after mesh repartitioning... only the
"n_threads>1" case is *likely* to cause breakage, but the "multiple
meshes" case was enough to cause multiple unit tests to fail.