Fix heisenbug in all_tri()
In debug or devel modes, this assert was breaking code in parallel.

First of all, new_bndry_elements.size() wasn't what we wanted to
check, because it could be zero on some processors while not on
others.

Secondly, because || is short-circuiting, we could end up in a
parallel-only function on only a subset of the comm's ranks, thereby
losing sync.