Add CheckpointIO unit test.
This verifies that we can read and write multiple checkpoint files
from a single processor.  After reading all the files, we verify that
total number of elements is correct.

Add files written by this test to the CLEANFILES list in Makefile.am,
also add a few others that had been overlooked.