Fix GetPot off-by-1 bug.
(Look at the the GetPot::search(const char* Option) function to follow
this discussion.)

Suppose we run a code with ./a.out --redirect-output, and when the
search() is called cursor==2, that is, it is already == argv.size().
Then:

.) OldCursor will be set to 1 in the first if-test.
.) The first for-loop will be skipped completely.
.) The second for-loop would (previously) loop while c < OldCursor.

Therefore if the searched term was in entry argv.size()-1, it would
erroneously not be found.  This patch fixes that issue.  We probably
never noticed this before because our GetPot object always had two
copies of all the command line arguments in it...