Loosen tolerance on straight-line check in Edge3::volume().
A user reported [0] a failure in Edge3::volume() for the otherwise
normal element:

x0 = (0.0041755731555024704, 0.02103, 0.)
x1 = (0.0042782212723039047, 0.02103, 0.)
x2 = (0.0042268963965901322, 0.02103, 0.)

Loosening the straight-line tolerance check is sufficient to fix this
particular case, but the failure was actually caused by passing a nan
to the std::log function.  In debug mode, we now verify that the
argument is valid before calling std::log.

[0]: https://groups.google.com/d/topic/moose-users/-8eC6TVwT5g/discussion