Fix issue with errno.h inclusion in a namespace.
After a recent PETSc update, I started getting a compile error:

> In file included from /usr/include/errno.h:23:0,
>                  from ./include/libmesh/netcdf.h:15,
>                  from ./include/libmesh/exodusII.h:46,
>                  from ./include/libmesh/exodusII_io_helper.h:54,
>                  from ../src/mesh/exodusII_io_helper.C:19:
> /opt/moose/gcc_5.2.0/lib/gcc/x86_64-apple-darwin15.0.0/5.2.0/include/mm_malloc.h: In function ‘void* _mm_malloc(size_t, size_t)’:
> /opt/moose/gcc_5.2.0/lib/gcc/x86_64-apple-darwin15.0.0/5.2.0/include/mm_malloc.h:39:7: error: ‘__error’ was not declared in this scope
>        errno = EINVAL;
>        ^
> /opt/moose/gcc_5.2.0/lib/gcc/x86_64-apple-darwin15.0.0/5.2.0/include/mm_malloc.h:39:7: note: suggested alternative:
> /usr/include/sys/errno.h:80:14: note:   ‘exII::__error’
>  extern int * __error(void);
>               ^

This appears to be related to, but not exactly the same problem, that
was fixed in 7169d43.  This time I tracked it down to the fact that
OSX defines the "errno" constant as a macro, i.e.

> extern int * __error(void);
> #define errno (*__error())

When errno.h is wrapped in a namespace, the error above is triggered,
since the compiler is looking for __error() in the enclosing
namespace.  A simple fix seems to be simply including <errno.h> before
exodusII.h, not wrapped in any namespace.  This fix is more complete
and more portable than the previous fix (adding an errno_t typedef)
so I'm going to update the configure test to do this as well.