Use the -fno-limit-debug-info flag in dbg/devel mode for clang.
This fixes an issue with clang/lldb [0] on Linux where the debugger
complains that:

class X has a member variable Y whose type is a forward declaration,
not a complete definition

There are two possible flags that can prevent this error/warning
message, -fno-limit-debug-info and -fstandalone-debug. It appears the
former has been around longer [1], so we are going with that one.
This flag is apparently already turned on for OSX, which might explain
why we haven't encountered it before.

[0]: https://llvm.org/bugs/show_bug.cgi?id=18454
[1]: https://reviews.llvm.org/D8347