Add '-Wunused -Wunused-parameter' for clang in all modes.
These need to be enabled in optimized mode to catch parameters which are
only used for debugging.  Such parameters need to be wrapped in the
libmesh_dbg_var() macro.

Note: previously we were using -Wunused in dbg and devel modes, but
that apparently does not imply -Wunused-parameter, at least in the
version of clang I'm using.  There is a fairly comprehensive list of
warnings at http://fuckingclangwarnings.com, but -Wunused is not
listed there. -Wunused *does* imply -Wunused-variable, however.

Refs idaholab/moose#8388.