If renumbering is not allowed, don't renumber.
This reverts the changes made in 845c28d7.  I'm not sure what the
purpose of those changes actually was... they still allowed both
elements and nodes to be renumbered if there were any NULL entries in
the _elements and _nodes vectors, which is inconsistent and confusing.
I'm sure this will expose other areas in the framework where a
contiguous node and element number is assumed...

Refs #938, #1216.